examples of pattern matching
Prolog pattern matching is based on unification, specifically the Martelli-Montanari Algorithm that matches values of the same position, binding variables on one side to a value at corresponding position on the other side. A simple example, the length/2 predicate. 

A red-black-tree is a binary tree where each internal node has a color attribute red or black. No red node can have a red child, and every path from the root to an empty node must contain the same number of black nodes. As a consequence, the tree is balanced, and must be re-balanced after an insertion.